requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
build==1.0.3
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==41.0.4
distlib==0.3.8
dnspython==2.7.0
dulwich==0.21.6
email_validator==2.2.0
fastapi==0.115.5
fastjsonschema==2.18.1
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==24.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mongomock==4.3.0
more-itertools==10.1.0
msgpack==1.0.7
narwhals==1.18.3
numpy==2.2.0
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pexpect==4.9.0
pillow==11.0.0
pipenv==2024.4.0
pkginfo==1.10.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
protobuf==5.29.1
ptyprocess==0.7.0
pyarrow==18.1.0
pycparser==2.21
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.10.1
pyproject_hooks==1.0.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
rapidfuzz==3.5.2
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
SecretStorage==3.3.3
sentinels==1.0.0
shellingham==1.5.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.3
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tomlkit==0.12.1
tornado==6.4.2
trove-classifiers==2023.9.19
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.0.6
uvicorn==0.32.1
virtualenv==20.27.1
watchdog==6.0.0
zipp==3.17.0
